Dynamic session room details
